added boilerplate for crud, plus minor adjustments
