+++ src/pages/profile.tsx
@@ -1,0 +1,15 @@
+import Header from "@/components/header";
+
+const ProfilePage = () => {
+  return (
+    <>
+      <Header app />
+
+      <main className='container mx-auto p-4 grid place-content-center min-h-screen'>
+        <h1>Feature coming soon!</h1>
+      </main>
+    </>
+  );
+};
+
+export default ProfilePage;
